Rename library to CoreLocationClient
