+++ Examples/LocationManager/Common/LocalSearchClient/Client.swift
@@ -2,21 +2,21 @@ import MapKit
 import Dependencies
 
 extension DependencyValues {
-    public var localSearchClient: LocalSearchClient {
-        get { self[LocalSearchClient.self] }
-        set { self[LocalSearchClient.self] = newValue }
-    }
+  public var localSearchClient: LocalSearchClient {
+    get { self[LocalSearchClient.self] }
+    set { self[LocalSearchClient.self] = newValue }
+  }
 }
 
 extension LocalSearchClient: TestDependencyKey {
-    public static let previewValue = Self.noop
-    public static let testValue = Self.failing
+  public static let previewValue = Self.noop
+  public static let testValue = Self.noop
 }
 
 extension LocalSearchClient {
-    public static let noop = Self(
-        search: { _ in try await Task.never() }
-    )
+  public static let noop = Self(
+    search: { _ in try await Task.never() }
+  )
 }
 
 public struct LocalSearchClient {
