+++ Examples/LocationManager/Common/LocalSearchClient/Live.swift
@@ -1,8 +1,5 @@
-import Combine
 import ComposableArchitecture
 import MapKit
-
-
 
 extension LocalSearchClient: DependencyKey {
     public static let liveValue = Self(
