+++ Sources/CoreLocationClient/Models/AccuracyAuthorization.swift
