+++ Sources/CoreLocationClient/Models/Beacon.swift
