+++ Sources/CoreLocationClient/Models/BeaconConstraint.swift
