+++ Sources/CoreLocationClient/Models/Heading.swift
