+++ Sources/CoreLocationClient/Models/Visit.swift
